requirements.txt
aiohttp
python-telegram-bot
yfinance
